feat: remove deprecated example files and update LSP compatibility modules

+++ crates/perl-parser/src/perl_critic.rs
@@ -258,7 +258,7 @@ impl CriticAnalyzer {
                 policy: v.policy.clone(),
                 description: v.description.clone(),
                 severity: v.severity as u8,
-                line: v.range.start.line,
+                line: v.range.start.line as usize,
             })
             .collect()
     }
